requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
pydantic-settings>=2.0.0
jinja2>=3.1.0
click>=8.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
lxml>=5.0.0
pyyaml>=6.0.0
